validateGoLinux: remove
What validateGoLinux does is check that we have a Go Linux binary.

It does that by checking if "GOOS=linux" is in "go version -m" output,
which is not working for Go 1.16. When, it falls back to using "file"
binary to check we have an ELF executable (which in most cases means
a Linux executable).

We have checked for both Go and ELF executable already, so this check is
redundant. Remove it.

(The other issue this commit solves is validateGoLinux is unnecessarily
called multiple times).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>