feat: ability to ignore individual errors
This allows to exclude specific known errors for specific files, meaning
that instead of skipping all checks for a particular file, we can only
skip some specific ones (or, rather, ignore their results).

The rules can be specified globally, or per-tag, or per-component,
or per-rpm.

TODO: convert configs to use this.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>