fix: log the entire configuration
Some fields are now missing from the log. Generally, it is not very
realistic to keep the Log function in sync with the configuration.

The solution is to use %+v which will print all the fields.

While at it, do not print git commit since we already do better than
that in main.go (right after calling config.Log()).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>