Use rpm name, not NVRA, in config
The motivation for this is, a version or a release change in an RPM
should not require changes in configuration.

Ultimately, this will also allow to have less rules.

Implementation details: we have to return both name and NVRA from
rpm.GetAllRPMs as, unfortunately, there is no reliable way to extract
the name from NVRA.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>