feat: ignore specific known errors per file
This allows to exclude specific known errors for specific files, meaning
that instead of skipping all checks for a particular file, we can only
skip some specific ones (or, rather, ignore specific errors for specific
files).

The rules to ignore errors can be specified globally, or per-tag, component,
or rpm.

The implementation is somewhat complicated because we have multiple set
of rules, plus in case of payload scan we do not know the rpm name
beforehand.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>