fix: merge configs with same keys
Because per-component, per-tag, and per-rpm ignore maps contains a
structure, when we get a map value, the structure is being copied.
So, we assign to a copy of it, and will need to assign a new structure.

The alternative would be to use a pointer as a map value, but this
breaks config.Log.

This fixes merge of parts of additional configuration with the key which
is already present in the main configuration (e.g. when both configs has
"[payload.xxx]" section).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>