feat: check for dups when merging configs
This commit adds checking for duplicates when merging configs, to ensure
that versioned config does not contain something that is already in the
main config.

For CI, we error out when a duplicate is found, as those configs are
embedded and they should be clean.

During runtime, we merely warn about duplicates, since it's not a deal
breaker. Here is an example warning that a user would get:

> using config file: config.toml
> adding rules from embedded config for 4.13
> main config [payload.operator-lifecycle-manager-container].filter_files already contains /usr/bin/cpb

NOTE that errors lists are not covered by this commit, and are to be
addressed separately.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>