feat: check for dups in IgnoreLists when merging configs
This commit adds checking for duplicates in ErrIgnoreLists when merging
configs, to ensure that versioned config does not contain something that
is already there in the main config.

For CI, we error out when a duplicate is found, as those configs are
embedded and they should be clean.

During runtime, we merely warn about duplicates, since it's not a deal
breaker. Here is an example warning that a user would get:

> main config [[rpm.glibc-common.ignore]].error=ErrNotDynLinked.files already contains /usr/sbin/build-locale-archive

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>