feat: add config validation
First, we check that all the paths are absolute and clean (otherwise the
tool won't work as expected), and error out if they aren't.

Second, we check that there are no overlaps in directory lists (e.g.
listing "/some" and "/some/dir" makes the second entry redundant), and
also that there are no redundant file entries (e.g. having "/some" in
dirs and "/some/file" in files). Such errors are considered warnings
during runtime, but we fail the test for those when checking embedded
configs in CI.

Note that this is in no way a comprehensive check, and it is not
catching all possible configuration issues. For example, we do not check
that per-component (per-tag, per-rpm, per-error) rules are redundant
given the global ignore exists, nor do we check for duplicate entries in
file and directory lists.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>