exclude a statically-compiled utility for OLM
In 4.15+, OLM will not require that users provide an `opm` binary with
their index image. In cases where they do not, OLM simply copies the
catalog data from their index image and uses it directly. In order to
copy this data, we do not want to assume anything about the image that
contains the content we need to copy, so we statically compile a utility
that helps us copy the data around, instead of assuming the image has
`sh` or `cp`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>