Pass component information to scan local check
Previously, we weren't passing in the `OpenshiftComponent` to the local
scan check. This meant that when you ran a local scan against the file
system of a container mounted locally, check-payload wouldn't honor the
exclusions implemented in the global configuration or the versioned
configuration files available in the check-payload binary.

This commit updates the scan local command to pass this information in,
and ports the component string argument to an `OpenshiftComponent`
struct, allowing us to reuse the existing exclusions. Ultimately making
it so `scan local` and `scan payload` behave similarly.

Fixes #193