Skip CGO checks for binaries without crypto
Go programs that do not import crypto do not need to link with
openssl, and therefore do not require CGO enabled.  This change
skips the CGO checks when stdlib crypto is not present.

The symtable load is factored out into a standalone function, and the
table is cached in the baton for validations to use independently.
This allows us to preserve the current intended validation order.